refactor(organisations): keep findPage query building private to each adapter (PAE-1813)
