+++ src/repositories/organisations/contract/find-page.contract.js
@@ -1,54 +1,9 @@
 import { beforeEach, describe, expect } from 'vitest'
-import { buildOrganisation } from './test-data.js'
-
-const buildOrgWithName = (name) => {
-  const base = buildOrganisation()
-  return {
-    ...base,
-    companyDetails: { ...base.companyDetails, name }
-  }
-}
+import { buildOrgWithCriteria, buildOrgWithName } from './test-data.js'
 
 const insertNamed = async (repository, names) => {
   for (const name of names) {
     await repository.insert(buildOrgWithName(name))
-  }
-}
-
-/**
- * Builds an organisation carrying known search criteria: the first registration
- * gets the given registration number and the first accreditation the given
- * accreditation number. Everything else is left as buildOrganisation makes it,
- * so the registration/accreditation links stay intact.
- *
- * @param {{
- *   name?: string,
- *   orgId?: number,
- *   registrationNumber?: string,
- *   accreditationNumber?: string
- * }} [overrides]
- * @returns {Omit<import('#domain/organisations/model.js').Organisation, 'status'>}
- */
-const buildOrgWithCriteria = ({
-  name = 'Criteria Ltd',
-  orgId,
-  registrationNumber,
-  accreditationNumber
-} = {}) => {
-  const base = buildOrganisation(orgId === undefined ? {} : { orgId })
-  return {
-    ...base,
-    companyDetails: { ...base.companyDetails, name },
-    registrations: base.registrations.map((registration, index) =>
-      index === 0 && registrationNumber !== undefined
-        ? { ...registration, registrationNumber }
-        : registration
-    ),
-    accreditations: base.accreditations.map((accreditation, index) =>
-      index === 0 && accreditationNumber !== undefined
-        ? { ...accreditation, accreditationNumber }
-        : accreditation
-    )
   }
 }
